Backup proof
rclone will not backup anything if nothing has changed since the
last backup. This is a very common situation for some of our S3
buckest.

The problem becomes: how do we check that the backup actually dit
run and work seemlessly? Today, we check that we have some files,
in the backup, that are from today, in which case, we consider
that the backup worked well. But this can't work if nothing has
changed in the source bucket.

Here we let the backup to insert a proof of backup file in the
source S3 to bakcup (debatable) and we're sure that, at least,
this file will always change for each backup. That way we can
keep our way of checking the backups by checking there are files
from today in the backup.